hardware_manager/requirements.txt
# file: hardware_manager/requirements.txt
pyserial
requests
python-dotenv